Code cleanup and refactor
